Sync OpenBSD patchset 164:
Tidy the target parsing code a bit and correct the behaviour so that as before
a string with no colon as a target window is first looked up as a window then
as a session, noted by Iain Morgan.

Also attempt to clarify the description of the target specification in the man
page.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@1388 303bd1e2-03bb-47f1-b221-2a0928954661