Sync OpenBSD patchset 197:
Add an additional heuristic to work out the current session when run from the
command line. The name of all slave ptys in the server is known, so if the
client was run on a tty, look for any sessions containing that tty and use the
most recently created.

This is more reliable than looking at $TMUX if windows have been moved or
linked between sessions.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@1430 303bd1e2-03bb-47f1-b221-2a0928954661