Sync OpenBSD patchset 202:
There aren't many client message types or code to handle them so get rid of
the lookup table and use a switch, merge the tiny handler functions into it,
and move the whole lot to client.c.

Also change client_msg_dispatch to consume as many messages as possible and
move the call to it to the right place so it checks for signals afterwards.

Prompted by suggestions from eric@.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@1435 303bd1e2-03bb-47f1-b221-2a0928954661