Sync OpenBSD patchset 284:
When moving up or down in copy mode, save the cursor position and size of the
last line with content (width != 0) and use it to determine if the cursor
should be at the end of the line. Fixes problem of the cursor always jumping to
the end of the line when scrolling past a blank line.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@1568 303bd1e2-03bb-47f1-b221-2a0928954661