Sync OpenBSD patchset 400:
When backspace is received at the beginning of a line and the previous line was
wrapped, move the cursor back up to the end of the previous line.

Another one of the forgotten persons requested this quite a while ago (I need
to start noting names on todo items...) when it was quite hard to
implement. Now it is easy and I don't see it can do any harm, so hey presto...


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@1740 303bd1e2-03bb-47f1-b221-2a0928954661