Sync OpenBSD patchset 426:
Some terminals don't correctly clear their let's-wrap flag after changing the
scroll region (which moves the cursor to 0,0). This means that if the cursor
was at the edge of the screen, any further output after scroll region change
incorrectly causes a line wrap. Add a workaround to move the cursor to position
0 if it is at the screen edge before changing scroll region.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@1772 303bd1e2-03bb-47f1-b221-2a0928954661