Sync OpenBSD patchset 544:
Permit top-bit-set characters to be entered in the status line. They could
already be set from the shell and are just passed through when printing (so
invisible characters or displaying on terminals with different character sets
may cause problems).

Note that entering UTF-8 may not work and in any case currently the status line
cannot display it correctly (outside of status-left/status-right).


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@1919 303bd1e2-03bb-47f1-b221-2a0928954661