Sync OpenBSD patchset 680:
Remove XXX comment and just close received fd if calloc() fails.

If this happens the imsg may no longer be usable as there may be queued
messages, but this is a) already the case with the code now, and b)
would be the case if recvmsg() fails anyway, so we can document that -1
from imsg_read() invalidates the struct imsgbuf.

discussed with and ok eric


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@2132 303bd1e2-03bb-47f1-b221-2a0928954661