Sync OpenBSD patchset 682:
If remain-on-exit is set, both the error callback and a SIGCHLD could
destroy the same pane (because the first one doesn't remove it from the
list of panes), causing the pane bufferevent to be freed twice. So don't
free it if the fd has already been set to -1, from Romain Francoise.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@2134 303bd1e2-03bb-47f1-b221-2a0928954661