Sync OpenBSD patchset 806:
Store sessions in an RB tree by name rather than a list, this is tidier
and allows them to easily be shown sorted in various lists
(list-sessions/choose-sessions).

Keep a session index which is used in a couple of places internally but
make it an ever-increasing number rather than filling in gaps with new
sessions.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux@2323 303bd1e2-03bb-47f1-b221-2a0928954661